Debounce state updates made by useInstallationsStatuses hook by 200ms
